refactor(desktop): update story imports to kebab-case filenames
